Correct recent Debian-only changes to catalog tutorial.
Many more remain, but it's getting unsightly to have both Debian and RPM
paths mentioned everywhere. The tutorial was much simpler before when it
only used one set of paths as an example and expected the user to look
up the other paths.  I'm not interested in making this Debian-only now,
though ...