I am wonderful! ;-)
Switch to extremely lightweight syntax for glossary elements, which
(besides cutting down the ammount of repeating XML text we'd have to
maintain), also allows us to include those glossary elements in-place
in other elements. (since they are now disassociated from the glossary context).

Changes made include very simple support for this in bin/generic-autogen,
and appropriate reduction in size on all glossary/* elements.

Minor fixes are done to some of the glossary elements as well, and WWWROOT
entry is deleted (it's windows-ism and I don't want that).

This might have as a consequence that *some* of glossary links temporarily
produce warnings (because the glossary filename was different than the
glossary id="" which we now auto-generate), but this doesn't break the
build process and will be fixed as less primary objective.