- Added support to include a preamble on autogenerated reference sets.
Also added a testing preamble for tags.xml. Including the whole introduction
to tags *really* doesn't look good (it's 15 pages of text, while all the
reference page is supposed to do is list the directives at the top of the
page!).

So in the preamble, instead of the whole text, I included a pointer to
a glossary entry. This is acceptable I think.